subscriptions, tags, ingredients

+++ backend/foodgram/api/views.py
@@ -5,12 +5,13 @@ from djoser.views import UserViewSet
 from rest_framework import filters, status
 from rest_framework.decorators import action
 from rest_framework.generics import get_object_or_404
+from rest_framework.pagination import PageNumberPagination
 from rest_framework.permissions import (AllowAny, IsAdminUser,
                                         IsAuthenticated, IsAuthenticatedOrReadOnly)
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet, ReadOnlyModelViewSet
 
-from api.serializers import (IngredientSerializer, RecipeSerializer,
+from api.serializers import (IngredientSerializer, RecipeSerializer, RecipeCreateSerializer,
                              SubscribeSerializer, TagSerializer, CustomUserSerializer)
 from recipes.models import Ingredient, Recipe, Tag
 
@@ -21,17 +22,7 @@ class CustomUserViewSet(UserViewSet):
     queryset = User.objects.all()
     serializer_class = CustomUserSerializer
     permission_classes = (IsAuthenticated,)
-    
-    
-    #@action(["get"], detail=False)
-    #def me(self, request, *args, **kwargs):
-    #    self.get_object = self.get_instance
-    #    if request.user.is_anonymous:
-    ##        return Response(
-    #           status=status.HTTP_401_UNAUTHORIZED
-    #        )
-    #    else:
-    #        return self.retrieve(request, *args, **kwargs)
+    pagination_class = PageNumberPagination
 
     @action(
         detail=True,
@@ -39,54 +30,59 @@ class CustomUserViewSet(UserViewSet):
         permission_classes=[IsAuthenticated]
     )
     def subscribe(self, request, id):
+        """Подписаться и отписаться от пользователя с известным id"""
         user = request.user
         author = get_object_or_404(User, id=id)
 
         if request.method == 'POST':
             if user.id == author.id:
                 return Response(
-                    'Нельзя подписаться на себя',
+                    {'errors':'Нельзя подписаться на себя'},
                     status=status.HTTP_400_BAD_REQUEST
                 )
             elif Subscription.objects.filter(user=user, author=author).exists():
                 return Response(
-                    'Вы уже подписаны на данного автора',
+                    {'errors':'Вы уже подписаны на данного автора'},
                     status=status.HTTP_400_BAD_REQUEST
                 )
             else:
                 subscription = Subscription.objects.create(
                     user=user, author=author
                 )
-                subscription.save()
-                serializer = SubscribeSerializer(IsAuthenticated,
+                serializer = SubscribeSerializer(author,
                                                  context={"request": request})
                 return Response(serializer.data, status=status.HTTP_201_CREATED)
 
         if request.method == 'DELETE':
-            subscription = get_object_or_404(Subscription,
-                                             user=user,
-                                             author=author)
-            subscription.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            subscription = Subscription.objects.filter(user=user, author=author)
+            if subscription.exists():
+                subscription.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response(
+                    {'errors': 'Вы не были подписаны на этого автора'},
+                    status=status.HTTP_400_BAD_REQUEST)
 
     @action(
         detail=False,
         permission_classes=[IsAuthenticated]
     )
     def subscriptions(self, request):
+        """Список подписок текущего авторизованного пользователя"""
         user = request.user
         queryset = User.objects.filter(subscribers__user=user)
-        pages = self.paginate_queryset(queryset)
-        serializer = SubscribeSerializer(pages,
-                                         many=True,
-                                         context={'request': request})
+        serializer = SubscribeSerializer(
+            self.paginate_queryset(queryset),
+            many=True,
+            context={'request': request}
+        )
         return self.get_paginated_response(serializer.data)
 
 
 class IngredientViewSet(ReadOnlyModelViewSet):
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
-    permission_classes = (IsAdminUser,)
+    permission_classes = (AllowAny,)
     filter_backends = (filters.SearchFilter,)
     search_fields = ('^name',)
 
@@ -95,7 +91,7 @@ class IngredientViewSet(ReadOnlyModelViewSet):
 class TagViewSet(ReadOnlyModelViewSet):
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
-    permission_classes = (IsAdminUser,)
+    permission_classes = (AllowAny,)
     filter_backends = (filters.SearchFilter,)
     search_fields = ('^name',)
 
@@ -105,5 +101,11 @@ class RecipeViewSet(ModelViewSet):
     serializer_class = RecipeSerializer
     permission_classes = (AllowAny,)
 
+    def perform_create(self, serializer):
+        serializer.save(author=self.request.user)
 
+    def get_serializer_class(self):
+        if self.action =='list' or self.action == 'retrieve':
+            return RecipeSerializer
+        return RecipeCreateSerializer
